CHANGE: make AppExeName from AppName, add build_deb to build-compose.yml

+++ build_deb.sh
@@ -1,28 +1,37 @@
 # user inputs
-AppName=mycalc
-AppExeName=MyCalc
-IconFile="icon.svg"
-Version=0.0.1
-Description="this is description"
-AppPublisher="muenai Co.,Ltd."
-AppPublisherEmail="service@muenai.com"
-AppUrl="https://www.muenai.com"
+AppName=$AppName
+#AppExeName=$AppExeName # executable name. The file path should be dist/AppName/AppExeName
+IconFile=$IconFile # path to icon file
+Version=$Version # version i.e. 0.2.1
+Description=$Description # app description
+AppPublisher=$AppPublisher # maintainer or company
+AppPublisherEmail=$AppPublisherEmail # email of maintainer or company
+AppUrl=$AppUrl # Homepage url
+
+echo Build parameters:
+echo "    AppName=$AppName"
+#echo "    AppExeName=$AppExeName"
+echo "    IconFile=$IconFile"
+echo "    Version=$Version"
+echo "    Description=$Description"
+echo "    AppPublisher=$AppPublisher"
+echo "    AppPublisherEmail=$AppPublisherEmail"
+echo "    AppUrl=$AppUrl"
 
 # derived parameters
+AppExeName=$AppName
 SOURCE=dist/$AppName
 ARCH=$(dpkg --print-architecture)
 WORKDIR=dist_wizard
+
 # tmp directory
 ROOTDIR=$WORKDIR/$AppName
 DEBIAN=$ROOTDIR/DEBIAN
-ICON_FNAME=$(echo $IconFile | rev | cut -d "/" -f 1 | rev)
 
 BIN=/usr/bin/$AppExeName
 LIB=/usr/lib/$AppName
-ICON=/usr/share/icons/hicolor/scalable/apps/$AppExeName-$ICON_FNAME
 DESKTOP=/usr/share/applications/$AppExeName.desktop
 
-SRC_ICON=$ROOTDIR$ICON
 SRC_BIN=$ROOTDIR$BIN
 SRC_LIB=$ROOTDIR$LIB
 SRC_DESKTOP=$ROOTDIR$DESKTOP
@@ -36,6 +45,7 @@ mkdir -p $ROOTDIR/usr/share/icons/hicolor/scalable/apps
 mkdir -p $DEBIAN
 
 # write desktop file
+echo Write desktop file
 echo "[Desktop Entry]
 Version=$Version
 Name=$AppName
@@ -48,19 +58,27 @@ Type=Application
 " > $SRC_DESKTOP
 
 # copy source files
+echo Copy source files
 if [[ -d "$SRC_LIB" ]]; then
     rm -rf $SRC_LIB
 fi
 cp -r $SOURCE $SRC_LIB
 
 # create relative soft link to /usr/bin/
+echo Create relative soft link to executable file
 if [[ -f "$SRC_BIN" ]]; then
     rm -f $SRC_BIN
 fi
 ln -r -s $SRC_LIB/$AppExeName $SRC_BIN
 
 # copy icon file
-cp -f $IconFile $SRC_ICON
+echo Copy icon file
+if [[ -n "$IconFile" && -f "$IconFile" ]]; then
+    ICON_FNAME=$(echo $IconFile | rev | cut -d "/" -f 1 | rev)
+    ICON=/usr/share/icons/hicolor/scalable/apps/$AppExeName-$ICON_FNAME
+    SRC_ICON=$ROOTDIR$ICON
+    cp -f $IconFile $SRC_ICON
+fi
 
 # write control file
 echo "Source: $AppName
